refactor(events): consolidate domain event classes into a single file, removing deprecated files and updating imports in webapi modules for improved structure and clarity 🗂️✨
