--- issue-solver/src/issue_solver/events/domain_event.py
@@ -1,16 +1,0 @@
-from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from datetime import datetime
-
-
-@dataclass(kw_only=True, frozen=True)
-class DomainEvent(ABC):
-    @property
-    @abstractmethod
-    def occurred_at(self) -> datetime:
-        pass
-
-    @property
-    @abstractmethod
-    def process_id(self) -> str:
-        pass
